add db and front (todo: fix it)

+++ core/src/main/java/org/example/reporters/MetricsReporter.java
@@ -3,5 +3,5 @@ package org.example.reporters;
 import java.lang.annotation.Annotation;
 
 public interface MetricsReporter<T extends Annotation> {
-    void report(T annotation, String methodName, double durationMs);
+    void report(T annotation, String methodName, long durationMs);
 }
